Integrate `info` fields into `system status`
Drop the separate top-level `container info` command and fold its richer
payload (client, host, resource counts, and system defaults) into
`container system status`. `system status` keeps its existing role and
exit-code contract: it still reports `unregistered`/`not running` and
exits non-zero in those cases, but when the daemon is up it now emits the
full client/host/server/paths/resources/defaults view in table or JSON.

This avoids a Docker-shaped command name and gives downstream tooling a
single Apple-native machine-readable surface, per PR review feedback.
